pdf-qa-chat/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pdfplumber==0.11.4
faiss-cpu==1.8.0.post1
openai==1.45.0
tiktoken==0.7.0
numpy==1.26.4
